Add profile service methods and comments
